+++ quickstart/csharp/uwp/keyword-recognizer/helloworld/App.xaml
@@ -1,0 +1,7 @@
+﻿<Application
+    x:Class="helloworld.App"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:helloworld">
+
+</Application>
