+++ quickstart/csharp/uwp/keyword-recognizer/helloworld/MainPage.xaml
@@ -1,0 +1,30 @@
+﻿<Page
+    x:Class="helloworld.MainPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:helloworld"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <!-- <StackPanel> -->
+        <StackPanel Orientation="Vertical" HorizontalAlignment="Center"  Margin="20,50,0,0" VerticalAlignment="Center" Width="800">
+            <Button x:Name="EnableMicrophoneButton" Content="Enable Microphone"  Margin="0,0,10,0" Click="EnableMicrophoneButton_Click" Height="35"/>
+            <Button x:Name="SpeechRecognitionButton" Content="Listen for Keyword" Margin="0,10,10,0" Click="SpeechRecognitionButton_Click" Height="35"/>
+            <Button x:Name="SaveToFileButton" Content="Save to File" Margin="0,10,10,0" Click="SaveToFileButton_Click" Height="35"/>
+            <StackPanel x:Name="StatusPanel" Orientation="Vertical" RelativePanel.AlignBottomWithPanel="True" RelativePanel.AlignRightWithPanel="True" RelativePanel.AlignLeftWithPanel="True">
+                <TextBlock x:Name="StatusLabel" Margin="0,10,10,0" TextWrapping="Wrap" Text="Status:" FontSize="20"/>
+                <Border x:Name="StatusBorder" Margin="0,0,0,0">
+                    <ScrollViewer VerticalScrollMode="Auto"  VerticalScrollBarVisibility="Auto" MaxHeight="200">
+                        <!-- Use LiveSetting to enable screen readers to announce the status update. -->
+                        <TextBlock x:Name="StatusBlock" FontWeight="Bold" AutomationProperties.LiveSetting="Assertive"
+                        MaxWidth="{Binding ElementName=Splitter, Path=ActualWidth}" Margin="10,10,10,20" TextWrapping="Wrap"  />
+                    </ScrollViewer>
+                </Border>
+            </StackPanel>
+        </StackPanel>
+        <!-- </StackPanel> -->
+    </Grid>
+</Page>
