--- quickstart/csharp/uwp/keyword-recognizer/helloworld/helloworld.sln
@@ -1,51 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30011.22
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "helloworld", "helloworld\helloworld.csproj", "{06F33717-6410-4BF2-A51D-AF976A1A9A55}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|ARM = Debug|ARM
-		Debug|ARM64 = Debug|ARM64
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|ARM = Release|ARM
-		Release|ARM64 = Release|ARM64
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|ARM.ActiveCfg = Debug|ARM
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|ARM.Build.0 = Debug|ARM
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|ARM.Deploy.0 = Debug|ARM
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|ARM64.ActiveCfg = Debug|ARM64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|ARM64.Build.0 = Debug|ARM64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|ARM64.Deploy.0 = Debug|ARM64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|x64.ActiveCfg = Debug|x64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|x64.Build.0 = Debug|x64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|x64.Deploy.0 = Debug|x64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|x86.ActiveCfg = Debug|x86
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|x86.Build.0 = Debug|x86
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Debug|x86.Deploy.0 = Debug|x86
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|ARM.ActiveCfg = Release|ARM
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|ARM.Build.0 = Release|ARM
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|ARM.Deploy.0 = Release|ARM
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|ARM64.ActiveCfg = Release|ARM64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|ARM64.Build.0 = Release|ARM64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|ARM64.Deploy.0 = Release|ARM64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|x64.ActiveCfg = Release|x64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|x64.Build.0 = Release|x64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|x64.Deploy.0 = Release|x64
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|x86.ActiveCfg = Release|x86
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|x86.Build.0 = Release|x86
-		{06F33717-6410-4BF2-A51D-AF976A1A9A55}.Release|x86.Deploy.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4BA4F34A-9E96-4A79-A054-D65B0565A64E}
-	EndGlobalSection
-EndGlobal
