--- quickstart/csharp/uwp/keyword-recognizer/helloworld/helloworld/App.xaml
@@ -1,7 +1,0 @@
-﻿<Application
-    x:Class="helloworld.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:helloworld">
-
-</Application>
